We're Getting Somewhere....
>nav done
>card layout done
>pseudo-sign in function added
>moved styles to CSS because it stopped being broken
>a LOT of layout work and resize work
primary index is the focus here.
Thankfully, that can be used as a template to make setting up the others take significantly less time.

Using Ringo Starr's "Yer Baby" as a placeholder for thumbnails until the json blob is set up.

The only views that are problematic are the Ipad view (font too small) and the galaxy fold (too narrow of a screen). The iPad could be a problem, but atm the Galaxy Fold will have too be disregarded.